refactor: modularize plotting
