+++ Data/Networks/example_multivoltage_adapted.py
