+++ core/constraints_common.py
